Move platform and runBlocking into common
