+++ kotest-common/api/kotest-common.api
@@ -19,6 +19,26 @@ public abstract interface annotation class io/kotest/common/JVMOnly : java/lang/
 public abstract interface annotation class io/kotest/common/KotestInternal : java/lang/annotation/Annotation {
 }
 
+public final class io/kotest/common/Platform : java/lang/Enum {
+	public static final field JS Lio/kotest/common/Platform;
+	public static final field JVM Lio/kotest/common/Platform;
+	public static final field Native Lio/kotest/common/Platform;
+	public static final field WasmJs Lio/kotest/common/Platform;
+	public static final field WasmWasi Lio/kotest/common/Platform;
+	public static fun getEntries ()Lkotlin/enums/EnumEntries;
+	public static fun valueOf (Ljava/lang/String;)Lio/kotest/common/Platform;
+	public static fun values ()[Lio/kotest/common/Platform;
+}
+
+public final class io/kotest/common/Platform_jvmKt {
+	public static final fun getPlatform ()Lio/kotest/common/Platform;
+}
+
+public final class io/kotest/common/RunBlocking_jvmKt {
+	public static final fun runBlocking (Lkotlin/jvm/functions/Function1;)Ljava/lang/Object;
+	public static final fun runPromise (Lkotlin/jvm/functions/Function1;)V
+}
+
 public abstract interface annotation class io/kotest/common/SoftDeprecated : java/lang/annotation/Annotation {
 	public abstract fun message ()Ljava/lang/String;
 }
