+++ kotest-common/src/commonMain/kotlin/io/kotest/common/runBlocking.kt
@@ -1,4 +1,4 @@
-package io.kotest.engine
+package io.kotest.common
 
 expect fun <T> runBlocking(f: suspend () -> T): T
