+++ kotest-common/src/jsMain/kotlin/io/kotest/common/platform.js.kt
@@ -1,3 +1,3 @@
-package io.kotest.core
+package io.kotest.common
 
 actual val platform: Platform = Platform.JS
