+++ kotest-common/src/jvmMain/kotlin/io/kotest/common/runBlocking.jvm.kt
@@ -1,4 +1,4 @@
-package io.kotest.engine
+package io.kotest.common
 
 actual fun <T> runBlocking(f: suspend () -> T): T = kotlinx.coroutines.runBlocking { f() }
