+++ kotest-common/src/wasmWasiMain/kotlin/io/kotest/common/platform.wasmJs.kt
@@ -1,3 +1,3 @@
-package io.kotest.core
+package io.kotest.common
 
 actual val platform: Platform = Platform.WasmWasi
