+++ kotest-common/src/wasmWasiMain/kotlin/io/kotest/common/runBlocking.wasmWasi.kt
@@ -1,4 +1,4 @@
-package io.kotest.engine
+package io.kotest.common
 
 actual fun runPromise(f: suspend () -> Unit) {
    error("Promise is only available on kotest/js")
