+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/TestEngine.kt
@@ -2,7 +2,7 @@ package io.kotest.engine
 
 import io.kotest.common.KotestInternal
 import io.kotest.core.Logger
-import io.kotest.core.Platform
+import io.kotest.common.Platform
 import io.kotest.core.config.AbstractProjectConfig
 import io.kotest.core.project.TestSuite
 import io.kotest.engine.extensions.ExtensionRegistry
