+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/spec/interceptor/SpecRefInterceptorPipeline.kt
@@ -1,8 +1,8 @@
 package io.kotest.engine.spec.interceptor
 
 import io.kotest.core.Logger
-import io.kotest.core.Platform
-import io.kotest.core.platform
+import io.kotest.common.Platform
+import io.kotest.common.platform
 import io.kotest.core.spec.SpecRef
 import io.kotest.core.test.TestCase
 import io.kotest.engine.test.TestResult
