+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/test/names/DefaultDisplayNameFormatter.kt
@@ -1,9 +1,9 @@
 package io.kotest.engine.test.names
 
-import io.kotest.core.Platform
+import io.kotest.common.Platform
 import io.kotest.core.annotation.DisplayName
 import io.kotest.core.names.TestNameCase
-import io.kotest.core.platform
+import io.kotest.common.platform
 import io.kotest.core.test.TestCase
 import io.kotest.engine.config.ProjectConfigResolver
 import io.kotest.engine.config.TestConfigResolver
