+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/extensions/test/BeforeAnyExceptionTest.kt
@@ -1,6 +1,6 @@
 package com.sksamuel.kotest.engine.extensions.test
 
-import io.kotest.core.Platform
+import io.kotest.common.Platform
 import io.kotest.core.annotation.EnabledIf
 import io.kotest.core.annotation.LinuxOnlyGithubCondition
 import io.kotest.core.spec.IsolationMode
