+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/RequiresPlatformInterceptorTest.kt
@@ -1,6 +1,6 @@
 package com.sksamuel.kotest.engine.spec
 
-import io.kotest.core.Platform
+import io.kotest.common.Platform
 import io.kotest.core.annotation.RequiresPlatform
 import io.kotest.core.spec.style.FunSpec
