--- kotest-framework/kotest-framework-engine/src/nativeMain/kotlin/io/kotest/engine/runBlocking.kt
@@ -1,7 +1,0 @@
-package io.kotest.engine
-
-actual fun <T> runBlocking(f: suspend () -> T): T = kotlinx.coroutines.runBlocking { f() }
-
-actual fun runPromise(f: suspend () -> Unit) {
-   error("Promise is only available on kotest/js")
-}
